Updated Ground/Wall Prefabs
Figured out the tilemap stuff--I'll explain once we get to the point where we really need/want to use it. I messed with the prefabs and such to get a platform and wall prefab ready with the proper box collider dimensions. Use Floor for a long floor, Platform for a small one to, say, hover in the air to reach an above area, and use Wall as, well, a wall. I adjusted the box colliders on the Ground/Platform/Floor prefabs to be slightly below the top part of the material so it looks like the character is walking on the floor rather than hovering over it.

**NOTE: I noticed that the walking animation is a bit off. It keeps thinking the character is jumping (grounded turns to notGrounded) when they are just walking straight in one direction, making them bounce almost constantly. Another thing is that if you press yourself up against a wall, you no longer fall down/can jump, move back to wall, jump again.