Maze Created
I created the maze area and the space above the maze. However, the area leading off to the upper-left needs a doorway and to be resized to fit the rest of the map/the open space around it. The vertical hall for the drop needs to be shortened and the horizontal hall at the bottom of the map needs to be finished.