Door Options and (Most of a) Title Screen
I added a few images for doors we can choose from and created a new scene for the title screen. It doesn't have any functionality yet, but basically the little light on the "START" button should light up when the user hovers the mouse over it and unlight when the mouse is removed. And, of course, it should start the game when clicked. The button itself is invisible, with the MenuOption prefab over it as the sprite. I was gonna add a cool animation where the title drops down from offscreen but ran out of time. We still need to add the sound. Things look weird in the Scene pane but look fine in the Game pane. I also changed the key from a square to the keycard and added in a cool font for the title screen.