Color/Timer Edits
The timer color turns from green to yellow when there are 30 seconds remaining, then from yellow to red when there are 10 seconds remaining. The timer also includes milliseconds now to emphasize the feeling of racing against the clock (and it looks cool). All menus now highlight the text in green and light up the light on the corresponding option when said option/button is hovered over.