Ignore Last Push, Look at This Instead
In the NewGraphics folder there are individual tile objects that can be dragged and dropped onto the tilemap and erased with the tilemap editor's eraser tool. Each tile counts as one 'square' on the tilemap regardless of its size, though larger tiles spread to fill the squares around them too. It's not going to be easy in the slightest to try and throw them all in a tile palette because they're not all a consistent size, so dragging and dropping instances of the tile onto the tilemap may work better. If all else fails, we can do exactly what we did with the rest of our map and just work with the sprites directly.