Started Graphics Updates
I redid the beginning hallway with the new sprites (stored under the GameObject "MapRedone") and got it to work with our existing physics. I deleted all the old tilemap files that don't/won't work so they're out of the way. I also deleted the old part of the map that I had remade so it was out of the way. I created a scene called "Instructions" that is currently just a duplicate of the "You Win" scene--eventually, it'll have a picture of the keycard the person is supposed to get with text saying "Find the keycard and escape the facility!" or something like that along with the movement instructions for a few seconds before the game actually starts. I'm going to work on everything more but wanted to push since I've made a lot of changes.