Added Credits, Good Ending, and Bad Ending
Once the game is ended, a short cutscene plays. If the player is successful, a cutscene of the ship flying away from the exploding space station plays before the credits play. Upon the credits being completed, the player is rerouted to the main menu screen. If the player is unsuccessful, a short cutscene of the station and spaceship blowing up plays before rerouting the player to the Game Over screen. The Instructions scene has music now as well.